Added backend logic and testing script with mock data
Added backendLogic.js which contains a main_handler function to 'dispatch' incoming requests for backend logic. includes placeholder implementations for all core backend methods (generateIdeas, queryHashtag, queryOpenAI, generateTrends, and semanticSearch) except for saving/discarding ideas produced by generateIdeas. Current implementation of generateIdeas and queryHashtags use mock data for testing, remaining are entirely placeholder due to OpenAI API not being integrated yet.

Added testBackend.js which is a testing script specifically for quick testing of functions in backendLogic.js.

Mockup tables (hashtags, keywords, sentimentAnalysis, posts, niches, and users) are artificially populated in backendLogic.js for testing purposes.

Currently all methods are confirmed to return an array or arrays of Strings, except for queryOpenAI which returns a single String (placeholder for the bot's respones) and queryHashtag which returns averageScore as the 'health' metric for the returned keyword. Ideally this will make it easier to 'plug in' the backend to the frontend button calls because 1. the outut format is known and 2. adjusting the logic of the placeholders shouldn't affect the format of the output at this point.